added hamburger button for responsive nav bar
This JavaScript function toggles the visibility of elements with the class "tooltip". It iterates through all elements with this class, and if an element's class is "tooltip", it adds the "hide" class to hide it; otherwise, it removes the "hide" class to show it. This effectively toggles the visibility of the elements.